+++ packages/Testing/src/Laravel/ApplicationInitiator.php
@@ -4,10 +4,12 @@ namespace Aedart\Testing\Laravel;
 
 use Aedart\Testing\Laravel\Bootstrap\LoadSpecifiedConfiguration;
 use Aedart\Testing\Laravel\Database\MigrateProcessor;
+use Aedart\Utils\Str;
 use Closure;
 use Illuminate\Container\Container;
 use Illuminate\Foundation\Application;
 use Illuminate\Support\Facades\Facade;
+use Illuminate\Support\ServiceProvider;
 use InvalidArgumentException;
 use Orchestra\Testbench\Concerns\Testing;
 use Orchestra\Testbench\Foundation\PackageManifest;
@@ -36,6 +38,23 @@ trait ApplicationInitiator
     protected string $environment = 'testing';
 
     /**
+     * When true, APP_KEY environment variable is
+     * generated and set.
+     *
+     * @var bool
+     */
+    protected bool $mustGenerateAppKey = false;
+
+    /**
+     * Automatically enables package discoveries.
+     *
+     * @see ignorePackageDiscoveriesFrom
+     *
+     * @var bool
+     */
+    protected bool $enablesPackageDiscoveries = false;
+
+    /**
      * Start the Laravel application
      *
      * <br />
@@ -52,8 +71,12 @@ trait ApplicationInitiator
             return $this;
         }
 
-        // Set the environment
+        // Set the environment and application key
         $this->setApplicationEnvironment($this->environment);
+        if ($this->mustGenerateAppKey) {
+            $key = $this->generateAppKey();
+            $this->setAppKeyEnvironmentVariable($key);
+        }
 
         // Setup test environment
         $this->setUpTheTestEnvironment();
@@ -79,6 +102,10 @@ trait ApplicationInitiator
 
         // Tear down test environment
         $this->tearDownTheTestEnvironment();
+
+        // Empty evt. paths to be published, or we risk that when a publish
+        // command is invoked, unintended resources are published.
+        ServiceProvider::$publishes = [];
 
         // Clear the service container's own instance, which
         // is set to Laravel's Application at this point.
@@ -167,6 +194,47 @@ trait ApplicationInitiator
         putenv('APP_ENV=' . $environment);
 
         return $this;
+    }
+
+    /**
+     * Sets the APP_KEY environment variable
+     *
+     * @param  string  $key
+     *
+     * @return self
+     */
+    protected function setAppKeyEnvironmentVariable(string $key): static
+    {
+        // Debug
+//        ConsoleDebugger::output('APP_KEY = ' . $key);
+
+        putenv('APP_KEY=' . $key);
+
+        return $this;
+    }
+
+    /**
+     * Set whether value for APP_KEY environment must be generated, or not
+     *
+     * @param  bool  $generate  [optional]
+     *
+     * @return self
+     */
+    public function mustGenerateAppKey(bool $generate = true): static
+    {
+        $this->mustGenerateAppKey = $generate;
+
+        return $this;
+    }
+
+    /**
+     * Generates a new application key
+     *
+     * @return string
+     */
+    protected function generateAppKey(): string
+    {
+        return Str::random(32);
     }
 
     /**